create sidebar modal with chats for user

+++ api/src/handlers/getRandomPosts.ts
@@ -51,8 +51,8 @@ export const handler: APIGatewayProxyHandlerV2 = async (event) => {
       CASE WHEN p.author IS NOT NULL THEN 'persona' ELSE 'user' END AS author_type,
       COALESCE(per.username, u.username) AS author_username,
       p.created_at,
-      COUNT(DISTINCT l.id) AS likes_count,
-      COUNT(DISTINCT c.id) AS comments_count
+      COUNT(DISTINCT l.id)::int AS likes_count,
+      COUNT(DISTINCT c.id)::int AS comments_count
     FROM posts p
     LEFT JOIN personas per ON p.author = per.persona_id
     LEFT JOIN users u ON p.user_author = u.id
